requirements.txt
asgiref==3.8.1
attrs==24.2.0
behave==1.2.6
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==5.1.4
django-crispy-forms==2.3
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
psycopg2-binary==2.9.10
PyHamcrest==2.1.0
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
